# client/configs/openapi/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenGraph API",
    "description": "AI/ML model and dataset Web3 blockchain infrastructure server",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/dictionary-categories/": {
      "post": {
        "tags": [
          "dictionary-categories"
        ],
        "summary": "Create Dictionary Category",
        "description": "Create a new dictionary-category association.",
        "operationId": "create_dictionary_category_api_v1_dictionary_categories__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DictionaryCategoryCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DictionaryCategoryRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/dictionary-categories/{dictionary_id}": {
      "get": {
        "tags": [
          "dictionary-categories"
        ],
        "summary": "Get Dictionary Categories",
        "description": "List all categories related to a dictionary.",
        "operationId": "get_dictionary_categories_api_v1_dictionary_categories__dictionary_id__get",
        "parameters": [
          {
            "name": "dictionary_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Dictionary Id"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 10,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/dictionary-categories/{dictionary_id}/{category_id}": {
      "delete": {
        "tags": [
          "dictionary-categories"
        ],
        "summary": "Delete Dictionary Category",
        "description": "Delete a dictionary-category association.",
        "operationId": "delete_dictionary_category_api_v1_dictionary_categories__dictionary_id___category_id__delete",
        "parameters": [
          {
            "name": "dictionary_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Dictionary Id"
            }
          },
          {
            "name": "category_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Category Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/dictionary-categories/batch": {
      "post": {
        "tags": [
          "dictionary-categories"
        ],
        "summary": "Create Dictionary Categories Batch",
        "description": "Create multiple dictionary-category associations for one dictionary.\nAll operations succeed or all fail (atomic transaction).",
        "operationId": "create_dictionary_categories_batch_api_v1_dictionary_categories_batch_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DictionaryCategoryBatchCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DictionaryCategoryRead"
                  },
                  "type": "array",
                  "title": "Response Create Dictionary Categories Batch Api V1 Dictionary Categories Batch Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CategoryListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/CategoryRead"
            },
            "type": "array",
            "title": "Items",
            "description": "Category list"
          },
          "total": {
            "type": "integer",
            "title": "Total",
            "description": "Total count"
          },
          "page": {
            "type": "integer",
            "title": "Page",
            "description": "Current page"
          },
          "limit": {
            "type": "integer",
            "title": "Limit",
            "description": "Page size"
          },
          "pages": {
            "type": "integer",
            "title": "Pages",
            "description": "Total pages"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "limit",
          "pages"
        ],
        "title": "CategoryListResponse",
        "description": "Category list response schema"
      },
      "CategoryRead": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name",
            "description": "Category Name"
          },
          "id": {
            "type": "integer",
            "title": "Id",
            "description": "Category ID"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "Creation timestamp"
          }
        },
        "type": "object",
        "required": [
          "name",
          "id",
          "created_at"
        ],
        "title": "CategoryRead"
      },
...
      "DictionaryCategoryBatchCreate": {
        "properties": {
          "dictionary_id": {
            "type": "integer",
            "title": "Dictionary Id",
            "description": "Dictionary ID"
          },
          "category_ids": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Category Ids",
            "description": "List of Category IDs"
          }
        },
        "type": "object",
        "required": [
          "dictionary_id",
          "category_ids"
        ],
        "title": "DictionaryCategoryBatchCreate",
        "description": "Batch dictionary-category association creation schema"
      },
      "DictionaryCategoryCreate": {
        "properties": {
          "dictionary_id": {
            "type": "integer",
            "title": "Dictionary Id",
            "description": "Dictionary ID"
          },
          "category_id": {
            "type": "integer",
            "title": "Category Id",
            "description": "Category ID"
          }
        },
        "type": "object",
        "required": [
          "dictionary_id",
          "category_id"
        ],
        "title": "DictionaryCategoryCreate",
        "description": "Dictionary-category association creation schema"
      },
      "DictionaryCategoryRead": {
        "properties": {
          "dictionary_id": {
            "type": "integer",
            "title": "Dictionary Id",
            "description": "Dictionary ID"
          },
          "category_id": {
            "type": "integer",
            "title": "Category Id",
            "description": "Category ID"
          }
        },
        "type": "object",
        "required": [
          "dictionary_id",
          "category_id"
        ],
        "title": "DictionaryCategoryRead",
        "description": "Dictionary-category association read schema"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}